# deployrunner/api/rest/docs/swagger.yaml
basePath: /api/deploy-installer/v1
definitions:
...
  api_rest_proton_component.HTTPError:
    properties:
      code:
        description: 错误分类下的错误码
        type: integer
      message:
        description: 具体错误信息
        type: string
      status:
        description: 状态码
        type: integer
    type: object
...
  api_rest_proton_component.MongoDBFull:
    properties:
      config:
        description: 系统级别配置,以字典对象表示。系统内应用与组件可见,优先级仅高于默认配置最低,会被其他层级配置覆盖
        type: object
      description:
        description: 描述
        type: string
      info:
        $ref: '#/definitions/types.MongoDBComponentInfo'
      instance:
        allOf:
        - $ref: '#/definitions/types.ComponentMongoDB'
        description: mq连接对象绑定的内建消息队列对象实例配置,详情见各类内建消息队列配置,类型见info字段中的MQType字段
      name:
        type: string
      namespace:
        description: 系统运行的k8s命名空间
        type: string
      sid:
        description: 系统ID
        type: integer
      systemName:
        description: 系统可视化名称,长度不大于20字符
        type: string
      type:
        type: string
    type: object
...
  api_rest_proton_component.OpensearchFull:
    properties:
      config:
        description: 系统级别配置,以字典对象表示。系统内应用与组件可见,优先级仅高于默认配置最低,会被其他层级配置覆盖
        type: object
      description:
        description: 描述
        type: string
      info:
        $ref: '#/definitions/types.OpensearchComponentInfo'
      instance:
        allOf:
        - $ref: '#/definitions/types.ComponentOpensearch'
        description: mq连接对象绑定的内建消息队列对象实例配置,详情见各类内建消息队列配置,类型见info字段中的MQType字段
      name:
        type: string
      namespace:
        description: 系统运行的k8s命名空间
        type: string
      sid:
        description: 系统ID
        type: integer
      systemName:
        description: 系统可视化名称,长度不大于20字符
        type: string
      type:
        type: string
    type: object
...
  components.MongoDBComponentInfo:
    properties:
      admin_key:
        type: string
      auth_source:
        type: string
      hosts:
        type: string
      mgmt_host:
        description: MgmtInfo
        type: string
      mgmt_port:
        type: integer
      options: {}
      password:
        type: string
      port:
        type: integer
      replica_set:
        type: string
      source_type:
        type: string
      ssl:
        type: boolean
      username:
        type: string
    type: object
  components.MongoDBComponentParams:
    properties:
      admin_passwd:
        type: string
      admin_secret_name:
        type: string
      admin_user:
        type: string
      data_path:
        type: string
      hosts:
        items:
          type: string
        type: array
      namespace:
        type: string
      password:
        type: string
      replica_count:
        type: integer
      resources:
        $ref: '#/definitions/components.Resources'
      storage_capacity:
        type: string
      storageClassName:
        type: string
      username:
        description: |-
          //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
          新增配置
        type: string
    required:
    - password
    - username
    type: object
...
  components.OpensearchComponentInfo:
    properties:
      distribution:
        type: string
      hosts:
        type: string
      password:
        type: string
      port:
        type: integer
      protocol:
        type: string
      source_type:
        type: string
      username:
        type: string
      version:
        type: string
    type: object
  components.OpensearchComponentParams:
    properties:
      config:
        $ref: '#/definitions/components.OpensearchConfigs'
      data_path:
        type: string
      exporter_resources:
        $ref: '#/definitions/components.Resources'
      extraValues:
        additionalProperties: true
        type: object
      hosts:
        items:
          type: string
        type: array
      mode:
        type: string
      namespace:
        type: string
      replica_count:
        type: integer
      resources:
        $ref: '#/definitions/components.Resources'
      settings:
        additionalProperties: true
        type: object
      storage_capacity:
        type: string
      storageClassName:
        type: string
    type: object
  components.OpensearchConfigs:
    properties:
      hanlpRemoteextDict:
        type: string
      hanlpRemoteextStopwords:
        type: string
      jvmOptions:
        type: string
    type: object
...
  components.ResourceRequirements:
    properties:
      cpu:
        type: string
      memory:
        type: string
    type: object
  components.Resources:
    properties:
      limits:
        $ref: '#/definitions/components.ResourceRequirements'
      requests:
        $ref: '#/definitions/components.ResourceRequirements'
    type: object
...
  types.ComponentMongoDB:
    properties:
      info:
        $ref: '#/definitions/components.MongoDBComponentInfo'
      name:
        type: string
      params:
        $ref: '#/definitions/components.MongoDBComponentParams'
      type:
        type: string
      version:
        type: string
    type: object
...
  types.ComponentOpensearch:
    properties:
      info:
        $ref: '#/definitions/components.OpensearchComponentInfo'
      name:
        type: string
      params:
        $ref: '#/definitions/components.OpensearchComponentParams'
      type:
        type: string
      version:
        type: string
    type: object
...
  types.MongoDBComponentInfo:
    properties:
      admin_key:
        type: string
      auth_source:
        type: string
      hosts:
        type: string
      mgmt_host:
        description: MgmtInfo
        type: string
      mgmt_port:
        type: integer
      options: {}
      password:
        type: string
      port:
        type: integer
      replica_set:
        type: string
      source_type:
        type: string
      ssl:
        type: boolean
      username:
        type: string
    type: object
  types.OpensearchComponentInfo:
    properties:
      distribution:
        type: string
      hosts:
        type: string
      password:
        type: string
      port:
        type: integer
      protocol:
        type: string
      source_type:
        type: string
      username:
        type: string
      version:
        type: string
    type: object
...
info:
  contact:
    email: tiga.gan@aishu.cn
    name: API Support
  description: 安装/更新执行器HTTP API接口文档
  title: 安装/更新执行器HTTP API文档
  version: 0.1.0
paths:
...
  /components/info/mongodb:
    put:
      consumes:
      - application/json
      description: 更新mongodb连接信息
      parameters:
      - description: mongodb
        in: body
        name: obj
        required: true
        schema:
          $ref: '#/definitions/api_rest_proton_component.MongoDBFull'
      produces:
      - application/json
      responses:
        "200":
          description: 返回为空
          schema:
            type: integer
        "400":
          description: 客户端请求参数错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "404":
          description: 对象不存在，不允许更新
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "500":
          description: 系统内部错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
      summary: 更新mongodb连接信息
      tags:
      - proton
      - connect
      - mongodb
  /components/info/mongodb/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: mongodb
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: mongodb
          schema:
            $ref: '#/definitions/api_rest_proton_component.MongoDBFull'
        "400":
          description: 客户端请求参数错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "404":
          description: 对象不存在
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "500":
          description: 系统内部错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
      summary: 获取mongodb类型示例信息
      tags:
      - proton
      - connect
      - mongodb
...
  /components/info/opensearch:
    put:
      consumes:
      - application/json
      description: 更新opensearch连接信息
      parameters:
      - description: opensearch
        in: body
        name: obj
        required: true
        schema:
          $ref: '#/definitions/api_rest_proton_component.OpensearchFull'
      produces:
      - application/json
      responses:
        "200":
          description: 返回为空
          schema:
            type: integer
        "400":
          description: 客户端请求参数错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "404":
          description: 对象不存在，不允许更新
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "500":
          description: 系统内部错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
      summary: 更新opensearch连接信息
      tags:
      - proton
      - connect
      - opensearch
  /components/info/opensearch/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: opensearch-master
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: opensearch
          schema:
            $ref: '#/definitions/api_rest_proton_component.OpensearchFull'
        "400":
          description: 客户端请求参数错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "404":
          description: 对象不存在
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "500":
          description: 系统内部错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
      summary: 获取opensearch类型示例信息
      tags:
      - proton
      - connect
      - opensearch
...
  /components/release/mongodb:
    put:
      consumes:
      - application/json
      description: 更新proton mongodb
      parameters:
      - description: zookeeper配置信息
        in: body
        name: obj
        required: true
        schema:
          $ref: '#/definitions/types.ComponentMongoDB'
      produces:
      - application/json
      responses:
        "200":
          description: 返回为空
          schema:
            type: integer
        "400":
          description: 客户端请求参数错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "404":
          description: 对象不存在，不允许更新
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "500":
          description: 系统内部错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
      summary: 更新proton mongodb
      tags:
      - proton
      - ProtonCompoment
      - mongodb
  /components/release/mongodb/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 实例名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 实例数据
          schema:
            $ref: '#/definitions/types.ComponentMongoDB'
        "400":
          description: 客户端请求参数错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "404":
          description: 对象不存在
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "500":
          description: 系统内部错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
      summary: 获取proton的mongodb实例
      tags:
      - proton
      - ProtonCompoment
      - mongodb
...
  /components/release/opensearch:
    put:
      consumes:
      - application/json
      description: 更新proton opensearch
      parameters:
      - description: zookeeper配置信息
        in: body
        name: obj
        required: true
        schema:
          $ref: '#/definitions/types.ComponentOpensearch'
      produces:
      - application/json
      responses:
        "200":
          description: 返回为空
          schema:
            type: integer
        "400":
          description: 客户端请求参数错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "404":
          description: 对象不存在，不允许更新
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "500":
          description: 系统内部错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
      summary: 更新proton opensearch
      tags:
      - proton
      - ProtonCompoment
      - opensearch
  /components/release/opensearch/{name}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 实例名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 实例数据
          schema:
            $ref: '#/definitions/types.ComponentOpensearch'
        "400":
          description: 客户端请求参数错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "404":
          description: 对象不存在
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
        "500":
          description: 系统内部错误
          schema:
            $ref: '#/definitions/api_rest_proton_component.HTTPError'
      summary: 获取proton的opensearch实例
      tags:
      - proton
      - ProtonCompoment
      - opensearch
...
swagger: "2.0"